refactor: replace resume app's kbarActions to libs directory
